backend done

+++ app/app.py
@@ -1,7 +1,9 @@
 from fastapi import FastAPI, HTTPException, Form, File, UploadFile, Depends
 from fastapi_users import fastapi_users
+from sqlalchemy.sql.functions import user
+
 from app.schemas import PostCreate, PostResponse, UserUpdate, UserRead, UserCreate
-from app.db import Post, create_db_tables, get_async_session
+from app.db import Post, create_db_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -38,6 +40,7 @@ text_posts = {
 @app.post("/upload")
 async def upload_file(file: UploadFile = File(...),
                       caption: str = Form(""),
+                      user: User = Depends(current_active_user),
                       session: AsyncSession = Depends(get_async_session)
 ):
     temp_file_path = None
@@ -58,6 +61,7 @@ async def upload_file(file: UploadFile = File(...),
             post = Post(
                 caption=caption,
                 url=upload_result.url,
+                user_id=user.id,
                 file_type="video" if upload_result.file_type.startswith("video") else "image",
                 file_name=upload_result.name,
             )
@@ -82,22 +86,31 @@ async def get_feed(
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
 
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
+
     posts_data = []
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_name": post.file_name,
                 "file_type": post.file_type,
                 "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": user_dict.get(post.user_id, "unknown"),
             }
         )
     return {"posts": posts_data}
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(post_id: str,
+                      user: User = Depends(current_active_user),
+                      session: AsyncSession = Depends(get_async_session)):
     try:
         post_uuid = uuid.UUID(post_id)
         result = await session.execute(select(Post).where(Post.id == post_uuid))
@@ -105,7 +118,8 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
 
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
-
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You don't have permission to delete this post")
         await session.delete(post)
         await session.commit()
